Customer Insertion, Dishes and DishesOrder Classes update

+++ DAL/DishDB.cs
@@ -49,7 +49,7 @@ namespace DAL
                                 dish.NAMEDISH = (string)dr["NAMEDISH"];
 
                             if (dr["PRICEDISH"] != null)
-                                dish.PRICEDISH = (int)dr["PRICEDISH"];
+                                dish.PRICEDISH = (decimal)dr["PRICEDISH"];
                         }
                     }
                 }
